reddit-scrapper/requirements.txt
google-genai==1.19.0
python-dotenv==1.0.1
supabase==2.15.2
sentistrength==0.0.9
google==3.0.0
praw==7.8.1
prawcore==2.4.0
pandas==2.3.0
uvicorn==0.34.0